moved from unsigned long long to uint64_t

+++ src/Metric.cxx
@@ -42,11 +42,11 @@ Metric::Metric(double value, const std::string& name, std::chrono::time_point<st
   mValue(value), mName(name), mTimestamp(timestamp)
 {}
 
-Metric::Metric(unsigned long long value, const std::string& name, std::chrono::time_point<std::chrono::system_clock> timestamp) :
+Metric::Metric(uint64_t value, const std::string& name, std::chrono::time_point<std::chrono::system_clock> timestamp) :
   mValue(value), mName(name), mTimestamp(timestamp)
 {}
 
-boost::variant< int, std::string, double, unsigned long long > Metric::getValue() const
+boost::variant< int, std::string, double, uint64_t > Metric::getValue() const
 {
   return mValue;
 }
